refactored markdown component
